+++ developer/src/kmc/test/test-project-build.ts
@@ -16,7 +16,8 @@ describe('BuildProject', function () {
       debug: false,
       warnDeprecatedCode: true,
     });
-    // 4 messages == starting build, build successful x 2
+    // 4 messages == starting build x 2, build successful x 2
+    // callbacks.printMessages();
     assert.equal(callbacks.messages.length, 4);
     assert.isTrue(result);
   });