+++ windows/src/engine/keyman32/globals.h
@@ -151,7 +151,6 @@ public:
 
   static BOOL get_debug_KeymanLog();
   static BOOL get_debug_ToConsole();
-  static BOOL get_CoreIntegration();
   static void LoadDebugSettings();
 };
 
@@ -215,8 +214,6 @@ typedef struct tagKEYMAN64THREADDATA
 	  KeymanUIDisabled,
 	  FInitialised,
     FInitialising;
-
-  char ForceFileName[MAX_PATH];
 
   DWORD ActiveKeymanID;